[2927] - Move suggestions action to new ProviderSuggestionsController
