+++ spec/requests/api/v3/providers/suggest_provider_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-describe "GET /suggest" do
+describe "GET /provider-suggestions" do
   let(:jsonapi_renderer) { JSONAPI::Serializable::Renderer.new }
   let(:organisation) { create(:organisation) }
   let(:provider1) { create(:provider, provider_name: "PROVIDER 1", organisations: [organisation]) }
@@ -17,7 +17,7 @@ describe "GET /suggest" do
   end
 
   it "searches for a particular provider for the user in the current recruitment cycle" do
-    get "/api/v3/providers/suggest?query=#{provider2.provider_name}"
+    get "/api/v3/provider-suggestions?query=#{provider2.provider_name}"
 
     expect(JSON.parse(response.body)["data"]).to match_array([
       {
@@ -32,13 +32,13 @@ describe "GET /suggest" do
   end
 
   it "searches for a provider that is not in the current recruitment cycle" do
-    get "/api/v3/providers/suggest?query=#{provider4.provider_name}"
+    get "/api/v3/provider-suggestions?query=#{provider4.provider_name}"
 
     expect(JSON.parse(response.body)["data"]).to match_array([])
   end
 
   it "searches for a partial provider in the current recruitment cycle" do
-    get "/api/v3/providers/suggest?query=#{provider2.provider_name[0..3]}"
+    get "/api/v3/provider-suggestions?query=#{provider2.provider_name[0..3]}"
 
     expect(JSON.parse(response.body)["data"]).to match_array([
       {
@@ -61,29 +61,29 @@ describe "GET /suggest" do
   end
 
   it "limits responses to a maximum of 5 items" do
-    10.times do
+    20.times do
       create(:provider, provider_name: "provider X", organisations: [organisation])
     end
 
-    get "/api/v3/providers/suggest?query=provider"
+    get "/api/v3/provider-suggestions?query=provider"
 
     expect(JSON.parse(response.body)["data"].length).to eq(5)
   end
 
   it "returns bad request if query is empty" do
-    get "/api/v3/providers/suggest"
+    get "/api/v3/provider-suggestions"
 
     expect(response.status).to eq(400)
   end
 
   it "returns bad request if query is too short" do
-    get "/api/v3/providers/suggest?query=#{provider2.provider_name[0, 2]}"
+    get "/api/v3/provider-suggestions?query=#{provider2.provider_name[0, 2]}"
 
     expect(response.status).to eq(400)
   end
 
   it "returns bad request if start of query is not alphanumeric" do
-    get "/api/v3/providers/suggest?query=%22%22%22%22"
+    get "/api/v3/provider-suggestions?query=%22%22%22%22"
 
     expect(response.status).to eq(400)
   end
